# DogAdoption/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API with centralized dog adoption information.",
        "title": "DogAdoptionApp REST HATEOAS API",
        "contact": {
            "name": "Gustav Näslund, gn222gq",
            "email": "gn222gq@student.lnu.se"
        },
        "version": "1.0"
    },
    "host": "https://cscloud7-113.lnu.se/dogadoption",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves detailed information about the authenticated user based on the provided credentials.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get authenticated user information",
                "responses": {
                    "200": {
                        "description": "Success, returns detailed information about the authenticated user",
                        "schema": {
                            "$ref": "#/definitions/userdto.UserDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized, if the user credentials do not match or are invalid",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a user identified by its unique ID, provided the requester has the necessary permissions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully"
                    },
                    "400": {
                        "description": "Bad Request, if the ID parameter format is incorrect",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized, if the requester does not have permission to delete the user",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found, if no user matches the provided ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/webhook": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves detailed information about a specific user webhook identified by its unique ID for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users/{id}/webhook"
                ],
                "summary": "Get a user webhook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, returns detailed information about the user webhook",
                        "schema": {
                            "$ref": "#/definitions/userwebhookdto.UserWebhookDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized, if the user credentials are invalid or do not grant access to the requested resource",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found, if no resource matches the provided ID or the webhook does not belong to the user",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates information for a specific user webhook identified by its unique ID for the authenticated user based on the provided data in JSON format. Secret must be minimum 12 characters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users/{id}/webhook"
                ],
                "summary": "Update a user webhook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Webhook Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userwebhookdto.UpdateUserWebhookDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, returns the updated user webhook information",
                        "schema": {
                            "$ref": "#/definitions/userwebhookdto.UserWebhookDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request, if the request data is incomplete or has invalid values",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized, if the user credentials do not match or are invalid",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found, if no matching webhook resource is found for the given ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a new webhook for the authenticated user based on the provided webhook data in JSON format. Secret must be minimum 12 characters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users/{id}/webhook"
                ],
                "summary": "Create a new user webhook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Webhook Data",
                        "name": "webhook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userwebhookdto.NewUserWebhookDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success, returns the newly created user webhook information",
                        "schema": {
                            "$ref": "#/definitions/userwebhookdto.UserWebhookDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request, if the JSON body cannot be parsed, mandatory fields are missing, or the webhook data is incomplete",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized, if the user credentials do not match or are invalid",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found, if the specified user does not exist",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a webhook for the authenticated user based on the provided webhook ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users/{id}/webhook"
                ],
                "summary": "Delete a webhook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Webhook ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Webhook deleted successfully"
                    },
                    "400": {
                        "description": "Bad Request - No webhook found or ID must be a number",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid user credentials",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - No webhook found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Something went wrong internally, try again later",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "model.WebhookAction": {
            "type": "string",
            "enum": [
                "new_dog_added"
            ],
            "x-enum-varnames": [
                "NEW_DOG_ADDED"
            ]
        },
...
        "userdto.UserDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "webhook": {
                    "$ref": "#/definitions/userwebhookdto.UserWebhookDTO"
                }
            }
        },
        "userwebhookdto.NewUserWebhookDTO": {
            "type": "object",
            "properties": {
                "client_secret": {
                    "type": "string"
                },
                "endpoint_url": {
                    "type": "string"
                },
                "webhook_actions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "userwebhookdto.UpdateUserWebhookDTO": {
            "type": "object",
            "properties": {
                "client_secret": {
                    "type": "string"
                },
                "endpoint_url": {
                    "type": "string"
                },
                "webhook_actions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "userwebhookdto.UserWebhookDTO": {
            "type": "object",
            "properties": {
                "endpoint_url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "webhook_actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.WebhookAction"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}